Building out logic
Here I have just copied the logic I used for my points per game predictor just to use it as a template and I'm going to edit it as needed. I also set up the routes and a new migration for my db as I need a different set of stats here which are currently stored as a CSV file.